GetInt - GetInt64

+++ handler/AcceptAnswerHandler.go
@@ -11,8 +11,8 @@ type AcceptAnswerHandler struct {
 }
 
 func (self *AcceptAnswerHandler) Get() {
-	aid, _ := self.GetInt(":aid")
-	qid, _ := self.GetInt(":qid")
+	aid, _ := self.GetInt64(":aid")
+	qid, _ := self.GetInt64(":qid")
 	uid, _ := self.GetSession("userid").(int64)
 	role, _ := self.GetSession("userrole").(int64)
 	if aid > 0 && qid > 0 && uid > 0 {
